added fixes and cleanups
