+++ library.properties
@@ -1,9 +1,9 @@
-name=HDC2080
-version=1.0
-author=Texas Instruments
-maintainer=TinkeringTech <tinkeringtech@gmail.com>
-sentence=A library to communicate and configure the Texas Instruments HDC2080 Temperature & Humidity Sensor
-paragraph=Uses I2C and the Arduino Wire Library to communicate with the HDC2080
+name=Lime Labs HDC2080
+version=1.0.1
+author=Lime Labs GmbH, Texas Instruments, TinkeringTech, @stickbreaker
+maintainer=Lime Labs GmbH <code@limelabs.io>
+sentence=An Arduino library to configure and communicate with the Texas Instruments HDC2080 temperature & humidity sensor
+paragraph=Uses I2C and the Arduino Wire library to communicate with the HDC2080
 category=Sensors
-url=
+url=https://github.com/lime-labs/HDC2080-Arduino
 architectures=*
